--- a/benches/archival_mmr.rs
+++ b/benches/archival_mmr.rs
@@ -159,7 +159,6 @@ mod mutate {
                     for (new_leaf, leaf_index) in
                         digests.iter().zip(leaf_index_of_mutated_leafs.iter())
                     {
-                        let _mp = ammr.prove_membership_async(*leaf_index).await.0;
                         ammr.mutate_leaf(*leaf_index, *new_leaf).await;
                     }
                     if persist {